move gather_batched
